Handle the case when all 6 receiver channels are connected to arduino pins.
This works much better because we can also read the gyro.sw. and the CH5
potentiometer states and use them for something, and we can more
accurately read the cyclic stick values, independently of the HOV.PIT.
pot, of the Idle Up mode and of the throttle state.

Unfortuntely it uses up an additional two cables and two input pins, one
of them being the arduino LED output.  Or is it possible to do this with
only every second channel connected? (so three pins only?)